[Refactor] 배치 job config 위치 변경
